Remove no longer needed ManagerT type parameter on helpers
