+++ src/UniGetUI.PackageEngine.Managers.Dotnet/Helpers/DotNetDetailsHelper.cs
