+++ src/UniGetUI.PackageEngine.Managers.PowerShell/Helpers/PowerShellDetailsHelper.cs
