+++ src/UniGetUI.PackageEngine.Managers.PowerShell7/Helpers/PowerShell7DetailsHelper.cs
