+++ src/UniGetUI.PackageEngine.Managers.Vcpkg/Helpers/VcpkgSourceHelper.cs
@@ -1,0 +1,39 @@
+using UniGetUI.Core.SettingsEngine;
+using UniGetUI.Core.Tools;
+using UniGetUI.PackageEngine.Classes.Manager;
+using UniGetUI.PackageEngine.Classes.Manager.Providers;
+using UniGetUI.PackageEngine.Enums;
+using UniGetUI.PackageEngine.Interfaces;
+using UniGetUI.PackageEngine.ManagerClasses.Manager;
+
+namespace UniGetUI.PackageEngine.Managers.VcpkgManager
+{
+    internal sealed class VcpkgSourceHelper : BaseSourceHelper
+    {
+        public VcpkgSourceHelper(Vcpkg manager) : base(manager) { }
+
+        public override string[] GetAddSourceParameters(IManagerSource source)
+            => throw new NotImplementedException();
+
+        public override string[] GetRemoveSourceParameters(IManagerSource source)
+            => throw new NotImplementedException();
+
+        protected override OperationVeredict _getAddSourceOperationVeredict(IManagerSource source, int ReturnCode, string[] Output)
+            => throw new NotImplementedException();
+
+        protected override OperationVeredict _getRemoveSourceOperationVeredict(IManagerSource source, int ReturnCode, string[] Output)
+            => throw new NotImplementedException();
+
+        protected override IEnumerable<IManagerSource> GetSources_UnSafe()
+        {
+            List<ManagerSource> Sources = [];
+
+			foreach (string Triplet in Vcpkg.GetSystemTriplets()) {
+				Sources.Add(new ManagerSource(Manager, Triplet, Vcpkg.URI_VCPKG_IO));
+			}
+
+            return Sources;
+        }
+
+    }
+}
