+++ src/UniGetUI.PackageEngine.Managers.WinGet/ClientHelpers/IWinGetManagerHelpers.cs
