+++ src/UniGetUI.PackageEngine.Managers.WinGet/ClientHelpers/NativePackageHandler.cs
