--- src/UniGetUI.PackageEngine.PackageManagerClasses/Manager/BaseProviders/BaseSourceHelper.cs
@@ -1,80 +1,0 @@
-using UniGetUI.Core.Classes;
-using UniGetUI.Core.Logging;
-using UniGetUI.PackageEngine.Enums;
-using UniGetUI.PackageEngine.Interfaces;
-using UniGetUI.PackageEngine.Interfaces.ManagerProviders;
-
-namespace UniGetUI.PackageEngine.Classes.Manager.Providers
-{
-    public abstract class BaseSourceHelper<ManagerT> : IMultiSourceHelper where ManagerT : IPackageManager
-    {
-        public ISourceFactory Factory { get; }
-        protected ManagerT Manager;
-
-        public BaseSourceHelper(ManagerT manager)
-        {
-            Manager = manager;
-            Factory = new SourceFactory(manager);
-        }
-
-        public abstract string[] GetAddSourceParameters(IManagerSource source);
-        public abstract string[] GetRemoveSourceParameters(IManagerSource source);
-        protected abstract OperationVeredict _getAddSourceOperationVeredict(IManagerSource source, int ReturnCode, string[] Output);
-        protected abstract OperationVeredict _getRemoveSourceOperationVeredict(IManagerSource source, int ReturnCode, string[] Output);
-
-        public OperationVeredict GetAddOperationVeredict(IManagerSource source, int ReturnCode, string[] Output)
-        {
-            TaskRecycler<IEnumerable<IManagerSource>>.RemoveFromCache(_getSources);
-            if (ReturnCode is 999 && Output.Last() == "Error: The operation was canceled by the user.")
-            {
-                Logger.Warn("Elevator [or GSudo] UAC prompt was canceled, not showing error message...");
-                return OperationVeredict.Canceled;
-            }
-            return _getAddSourceOperationVeredict(source, ReturnCode, Output);
-        }
-
-        public OperationVeredict GetRemoveOperationVeredict(IManagerSource source, int ReturnCode, string[] Output)
-        {
-            TaskRecycler<IEnumerable<IManagerSource>>.RemoveFromCache(_getSources);
-            if (ReturnCode is 999 && Output.Last() == "Error: The operation was canceled by the user.")
-            {
-                Logger.Warn("Elevator [or GSudo] UAC prompt was canceled, not showing error message...");
-                return OperationVeredict.Canceled;
-            }
-            return _getRemoveSourceOperationVeredict(source, ReturnCode, Output);
-        }
-
-        /// <summary>
-        /// Loads the sources for the manager. This method SHOULD NOT handle exceptions
-        /// </summary>
-        protected abstract IEnumerable<IManagerSource> GetSources_UnSafe();
-
-        public virtual IEnumerable<IManagerSource> GetSources()
-            => TaskRecycler<IEnumerable<IManagerSource>>.RunOrAttach(_getSources, 15);
-
-        public virtual IEnumerable<IManagerSource> _getSources()
-        {
-            if (!Manager.IsReady()) { Logger.Warn($"Manager {Manager.Name} is disabled but yet GetSources was called"); return []; }
-
-            try
-            {
-                IEnumerable<IManagerSource> sources = GetSources_UnSafe().ToArray();
-                Factory.Reset();
-
-                foreach (IManagerSource source in sources)
-                    Factory.AddSource(source);
-
-                Logger.Debug($"Loaded {sources.Count()} sources for manager {Manager.Name}");
-                return sources;
-            }
-            catch (Exception e)
-            {
-                Logger.Error("Error finding sources for manager " + Manager.Name);
-                Logger.Error(e);
-                return [];
-            }
-
-
-        }
-    }
-}
